Remove some hand written java DAO implementations in favor of generated impls.
