+++ src/foam/dao/DatastoreDAO.java
@@ -24,26 +24,10 @@ import java.util.List;
 public class DatastoreDAO
   extends AbstractDAO
 {
-    protected ClassInfo        of_         = null;
-    protected PropertyInfo     primaryKey_ = null;
     protected DatastoreService datastore_  = DatastoreServiceFactory.getDatastoreService();
 
     protected DatastoreService getData() {
       return datastore_;
-    }
-
-    public ClassInfo getOf() {
-      return of_;
-    }
-
-    public DatastoreDAO setOf(ClassInfo of) {
-      of_ = of;
-      primaryKey_ = (PropertyInfo) of.getAxiomByName("id");
-      return this;
-    }
-
-    public PropertyInfo getPrimaryKey() {
-      return primaryKey_;
     }
 
     protected List<PropertyInfo> getProperties(ClassInfo c) {
