+++ src/foam/dao/JDAO.java
@@ -49,7 +49,7 @@ public class JDAO
   public JDAO(foam.core.X x, ClassInfo classInfo, String filename)
       throws IOException
   {
-    this(x, new MapDAO().setOf(classInfo), filename);
+    this(x, new MapDAO(classInfo), filename);
   }
 
   public JDAO(foam.core.X x, DAO delegate, String filename) throws IOException {
