+++ src/foam/java/refinements.js
@@ -276,6 +276,11 @@ foam.CLASS({
       class: 'Boolean',
       name: 'javaSupport',
       value: true
+    },
+    {
+      class: 'Boolean',
+      name: 'javaStatic',
+      value: false
     }
   ],
 
@@ -288,7 +293,7 @@ foam.CLASS({
         name: this.name,
         type: this.javaReturns || 'void',
         visibility: 'public',
-        static: this.isStatic(),
+        static: this.javaStatic,
         synchronized: this.synchronized,
         throws: this.javaThrows,
         args: this.args && this.args.map(function(a) {
@@ -300,7 +305,6 @@ foam.CLASS({
         body: this.javaCode ? this.javaCode : ''
       });
     },
-    function isStatic() { return false; }
   ]
 });
